Added autorole module
A module that automatically add roles whenever a user sends a message with a cooldown to avoid spending too much resources on it.
This can be used as a mean to check whether a member has a verified account as having a role bypasses the requirement for having a verified account.
Auto roles may be roles such as category roles or something like a member role.